Pulled more functions out into separate files
