+++ app/time/core/utils.js
@@ -335,7 +335,7 @@ export function Events() {
   return {
     event
   };
-};
+}
 
 export function pickInterval(range) {
   const intervals = [
@@ -348,278 +348,6 @@ export function pickInterval(range) {
     moment.duration(1, "years").asMilliseconds()
   ];
   return intervals[Math.max(exports.find(intervals, range.width()) - 1, 0)];
-};
-
-export function stringToMoment(date, format){
-  return moment(date, format);
-}
-
-/**
- * Read an iso duration into a moment.js object.
- * @param {string} duration
- * @returns {object} with moment.js info
- */
-export function isoDurationToMoment(duration) {
-  if (duration.charAt(0) !== "P") {
-    throw new Error(`expected P as starting duration : ${  duration}`);
-  }
-  const pattern = /(\d+)(\w)/g;
-  let date = null, time = null;
-  const values = {};
-
-  duration = duration.substring(1);
-  if (duration.indexOf("T") >= 0) {
-    const parts = duration.split("T");
-    date = parts[0];
-    time = parts[1];
-  } else {
-    date = duration;
-  }
-  const mapping = {
-    "Y": "years",
-    "M": "months",
-    "W": "weeks",
-    "D": "days",
-    "H": "hours",
-    "m": "minutes",
-    "S": "seconds"
-  };
-  function parse(chunk, time) {
-    function read(amount, part) {
-      if (time && part === "M") {
-        part = "m";
-      }
-      const mappedTo = mapping[part];
-      if (typeof mappedTo === "undefined") {
-        throw Error(`unknown duration specifier : ${  part}`);
-      }
-      values[mappedTo] = parseFloat(amount);
-    }
-    let next = pattern.exec(chunk);
-    while (next !== null) {
-      read(next[1], next[2]);
-      next = pattern.exec(chunk);
-    }
-  }
-  if (date !== null) {
-    parse(date, false);
-  }
-  if (time !== null) {
-    parse(time, true);
-  }
-  return values;
-}
-
-/**
- * Get a function for the provided duration that computes a new timestamp based on a
- * provided date and optional multiplier (negative for reverse).
- * @param {string} iso duration
- * @returns {function} offsetter(timestamp, multiplier=1)
- */
-export function createOffsetter(intervalOrDuration) {
-  const duration = typeof intervalOrDuration === "string" ? intervalOrDuration: intervalOrDuration.duration;
-  const values = exports.isoDurationToMoment(duration);
-  // as of writing, moment assumes y=365d and m=30d resulting in slow
-  // day of month shifts that break ticks from matching
-  // so we take care of this using a more accurate approach
-  // ** the current approach breaks down if the day of month is greater than
-  // 28 and day of month will no longer be retained (will shift)
-  if ("years" in values || "months" in values) {
-    const years = values.years;
-    const months = values.months;
-    values.years = 0;
-    values.months = 0;
-    const millis = moment.duration(values).asMilliseconds();
-    return (ts, mult) => {
-      mult = mult || 1;
-      const d = new Date(ts);
-      /* jshint eqnull:true */
-      let y = d.getUTCFullYear();
-      if (years != null) {
-        y += mult * years;
-      }
-      let m = d.getUTCMonth();
-      if (months != null) {
-        m += mult * months;
-      }
-      d.setUTCFullYear(y, m);
-      return d.getTime() + (mult * millis);
-    };
-  }
-  const offset = moment.duration(values).asMilliseconds();
-  return (ts, mult) => {
-    mult = mult || 1;
-    return ts + (mult * offset);
-  };
-
 }
 
 
-/**
- * Contains implementations of Date.parse and date.toISOString that match the
- *     ECMAScript 5 specification for parsing RFC 3339 dates.
- *     http://tools.ietf.org/html/rfc3339
- */
-
-
-
-
-//  Secure Hash Algorithm (SHA1)
-//  http://www.webtoolkit.info/
-
-export function sha1(msg) {
-  /* eslint-disable no-bitwise */
-  const rotate_left = (n, s) => (n << s) | (n >>> (32 - s));
-
-  const cvt_hex = (val) => {
-    let str = "";
-    let i;
-    let v;
-
-    for (i = 7; i >= 0; i--) {
-      v = (val >>> (i * 4)) & 0x0f;
-      str += v.toString(16);
-    }
-    return str;
-  };
-
-
-  const utf8Encode = (string) => {
-    string = string.replace(/\r\n/g, "\n");
-    let utftext = "";
-
-    for (let n = 0; n < string.length; n++) {
-
-      const c = string.charCodeAt(n);
-
-      if (c < 128) {
-        utftext += String.fromCharCode(c);
-      } else if ((c > 127) && (c < 2048)) {
-        utftext += String.fromCharCode((c >> 6) | 192);
-        utftext += String.fromCharCode((c & 63) | 128);
-      } else {
-        utftext += String.fromCharCode((c >> 12) | 224);
-        utftext += String.fromCharCode(((c >> 6) & 63) | 128);
-        utftext += String.fromCharCode((c & 63) | 128);
-      }
-
-    }
-    return utftext;
-  };
-
-  let blockstart;
-  let i, j;
-  const W = new Array(80);
-  let H0 = 0x67452301;
-  let H1 = 0xEFCDAB89;
-  let H2 = 0x98BADCFE;
-  let H3 = 0x10325476;
-  let H4 = 0xC3D2E1F0;
-  let A, B, C, D, E;
-  let temp;
-
-  msg = utf8Encode(msg);
-
-  const msg_len = msg.length;
-
-  const word_array = [];
-  for (i = 0; i < msg_len - 3; i += 4) {
-    j = msg.charCodeAt(i) << 24 | msg.charCodeAt(i + 1) << 16 |
-      msg.charCodeAt(i + 2) << 8 | msg.charCodeAt(i + 3);
-    word_array.push(j);
-  }
-
-  switch (msg_len % 4) {
-  case 0:
-    i = 0x080000000;
-    break;
-  case 1:
-    i = msg.charCodeAt(msg_len - 1) << 24 | 0x0800000;
-    break;
-
-  case 2:
-    i = msg.charCodeAt(msg_len - 2) << 24 | msg.charCodeAt(msg_len - 1) << 16 | 0x08000;
-    break;
-
-  case 3:
-    i = msg.charCodeAt(msg_len - 3) << 24 | msg.charCodeAt(msg_len - 2) << 16 |
-        msg.charCodeAt(msg_len - 1) << 8 | 0x80;
-    break;
-  default:
-    break
-  }
-
-  word_array.push(i);
-
-  while ((word_array.length % 16) !== 14) {
-    word_array.push(0);
-  }
-
-  word_array.push(msg_len >>> 29);
-  word_array.push((msg_len << 3) & 0x0ffffffff);
-
-
-  for (blockstart = 0; blockstart < word_array.length; blockstart += 16) {
-
-    for (i = 0; i < 16; i++) {
-      W[i] = word_array[blockstart + i];
-    }
-    for (i = 16; i <= 79; i++) {
-      W[i] = rotate_left(W[i - 3] ^ W[i - 8] ^ W[i - 14] ^ W[i - 16], 1);
-    }
-
-    A = H0;
-    B = H1;
-    C = H2;
-    D = H3;
-    E = H4;
-
-    for (i = 0; i <= 19; i++) {
-      temp = (rotate_left(A, 5) + ((B & C) | (~B & D)) + E + W[i] + 0x5A827999) & 0x0ffffffff;
-      E = D;
-      D = C;
-      C = rotate_left(B, 30);
-      B = A;
-      A = temp;
-    }
-
-    for (i = 20; i <= 39; i++) {
-      temp = (rotate_left(A, 5) + (B ^ C ^ D) + E + W[i] + 0x6ED9EBA1) & 0x0ffffffff;
-      E = D;
-      D = C;
-      C = rotate_left(B, 30);
-      B = A;
-      A = temp;
-    }
-
-    for (i = 40; i <= 59; i++) {
-      temp = (rotate_left(A, 5) + ((B & C) | (B & D) | (C & D)) + E + W[i] + 0x8F1BBCDC) & 0x0ffffffff;
-      E = D;
-      D = C;
-      C = rotate_left(B, 30);
-      B = A;
-      A = temp;
-    }
-
-    for (i = 60; i <= 79; i++) {
-      temp = (rotate_left(A, 5) + (B ^ C ^ D) + E + W[i] + 0xCA62C1D6) & 0x0ffffffff;
-      E = D;
-      D = C;
-      C = rotate_left(B, 30);
-      B = A;
-      A = temp;
-    }
-
-    H0 = (H0 + A) & 0x0ffffffff;
-    H1 = (H1 + B) & 0x0ffffffff;
-    H2 = (H2 + C) & 0x0ffffffff;
-    H3 = (H3 + D) & 0x0ffffffff;
-    H4 = (H4 + E) & 0x0ffffffff;
-
-  }
-
-  const localtemp = cvt_hex(H0) + cvt_hex(H1) + cvt_hex(H2) + cvt_hex(H3) + cvt_hex(H4);
-
-  return localtemp.toLowerCase();
-  /* eslint-enable no-bitwise */
-}
